backend/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework==3.13.1
sqlparse==0.4.2
mysqlclient==2.1.0